Move init_db to init_setup
